Make db_charmer_connection_proxy/level thread safe.
db_charmer_connection_proxy and _level were saved in class variables.
Proxy contains the real connections, so they are shared across all
threads.

The bug was found because it caused intermittent failures with
cucumber+phantomjs specs. This setup runs a webapp in a thread and runs
the cucumber specs in a different threads and it causes a race
condition.

There a case could made for making force_slave_reads thread safe too.